Async config validation
